src/requirements.txt
Django==1.7.5
djangorestframework==3.1.0
